+++ javascript/src/local/index.ts
@@ -1,4 +1,5 @@
 export * from "./core.js";
 export * from "./context.js";
+export * from "./pause.js";
 export * from "./shell.js";
 export * from "./tools.js";